% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/analyzeTS.R
\name{analyzeTS}
\alias{analyzeTS}
\title{Analyze time series with different method.}
\usage{
analyzeTS(TS, method, omitNA = TRUE, output = "data", ...)
}
\arguments{
\item{TS}{A time series with first column the Date.}

\item{method}{A string showing the method used to calculate mean value, e.g., "annual".
more information please refer to details.}

\item{omitNA}{A boolean showing in the calculation, whether NA is omitted, default is TRUE.}

\item{output}{A character representing the output type, if you want to use \code{analyzeTS_comb}
to combine different output from this function, you can choose \code{output = 'ggplot'}, if not
assigned, normal data array will be returned.}

\item{...,}{\code{title, x, y} showing the title and x and y axis of the plot, shoud be a string.}
}
\value{
The annual/monthly/seasonal analysis of the input time series.
}
\description{
Analyze time series with different method.
}
\details{
There are following methods to be selected,
"annual": annual rainfall of each year is plotted.
"winter", "spring", "autumn", "summer": seasonal rainfall of each year is plotted.
Month(number 1 to 12): month rainfall of each year is plotted, e.g. march rainfall of each year.

Since "winter" is a crossing year, 12, 1, 2, 12 is in former year, and 1, 2 are in latter year.
so winter belongs to the latter year.
}
\examples{
data(testdl)
TS  <- testdl[[1]]

# Get the mean spring precipitation.
a <- analyzeTS(TS, method = 'spring')
a <- analyzeTS(TS, method = 'spring', omitNA = FALSE)


a <- analyzeTS(TS, method = 'winter')
a
# Get special month precipitation, e.g. march.
a <- analyzeTS(TS, method = 3)
a

# We can also get annual precipitation.
a <- analyzeTS(TS, method = 'annual')
a
}

