% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/analyzeTS.R
\name{analyzeTS_comb}
\alias{analyzeTS_comb}
\title{Combine rainfall analysis together, to generate multiplot.}
\usage{
analyzeTS_comb(..., list = NULL, nrow = 1)
}
\arguments{
\item{...}{different barplots generated by \code{analyzeTS(, output = 'ggplot')}, refer to details.}

\item{list}{If input is a list containing different ggplot data, use l\code{list = inputlist}.
NOTE: yOU HAVE TO PUT A \code{list = }, before your list.}

\item{nrow}{A number showing the number of rows.}
}
\value{
A combined barplot of the analysis.
}
\description{
Combine rainfall analysis together, to generate multiplot.
}
\details{
..., representing different ouput generated by \code{analyzeTS(, output = 'ggplot')}, they
have to be of the same type, e.g.,
1. Jan precipitation of different years, Feb precipitation of different years, and...
They are both monthly precipitation, and they share x axis.

2. Mean monthly precipitation of different dataset. e.g., long term mean monthly precipitation
and short term mean monthly precipitation. They are both mean monthly precipitation.
}
\examples{
data(testdl)
TS  <- testdl[[1]]

#output type of analyzeTS() has to be 'ggplot'.
a <- analyzeTS(TS, method = 'spring', output = 'ggplot')
b <- analyzeTS(TS, method = 'summer', output = 'ggplot')

analyzeTS_comb(a, b)



a <- analyzeTS(TS, method = 3, output = 'ggplot')
b <- analyzeTS(TS, method = 4, output = 'ggplot')
analyzeTS_comb(a, b)
}

