% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/biasCorrect.R
\name{biasCorrect}
\alias{biasCorrect}
\title{Biascorrect the input timeseries or hyfo dataset}
\usage{
biasCorrect(frc, hindcast, obs, method = "delta", scaleType = "multi",
  input = "TS", preci = FALSE)
}
\arguments{
\item{frc}{a hyfo grid data output or a dataframe (time series) consists of Date column and one or more value columns,
representing the forecast to be calibrated.}

\item{hindcast}{a hyfo grid data output or a dataframe(time series) consists of Date column and one or more value columns,
representing the hindcast data. This data will be used in the calibration of the forecast, so it's better to have the same date period as
observation data. Check details for more information.}

\item{obs}{a hyfo grid data output or a dataframe (time series) consists of Date column and one or more value columns,
representing the observation data.}

\item{method}{bias correct method, including 'delta', 'scaling'...}

\item{scaleType}{only when the method "scaling" is chosen, scaleType will be available. Two different types
of scaling method, 'add' and 'mult', which means additive and multiplicative scaling method. More info check
details.}

\item{input}{If input is a time series, \code{input = 'TS'} needs to be assigned, or hyfo will take it as
an hyfo output grid file. Default is time series input, where in most of the cases we prefer. If your input
is a hyfo output file, \code{input = 'hyfo'}.}

\item{preci}{If the precipitation is biascorrected, then you have to assign \code{preci = TRUE}. Since for
precipitation, some biascorrect methods may not apply to, or some methods are specially for precipitation.
Default is FALSE.}
}
\description{
Biascorrect the input time series or dataset, the input time series or dataset should consist of observation, hindcast, and forecast.
observation and hindcast should belong to the same period, in order to calibrate. Then the modified forecast
will be returned. If the input is a time series, first column should be date column and rest columns should be
the value column. If the input is a hyfo dataset, the dataset should be the result of \code{loadNcdf}, or a list
file with the same format.
}
\details{
Since climate forecast is based on global condition, when downscaling to different regions, it may include
some bias, biascorrection is used then to fix the bias.

\strong{Hindcast}

In order to bias correct, we need to pick up some
data from the forecast to train with the observation, which is called hindcast in this function. Hindcast
should have \strong{EVERY} attributes that forecast has.

Hindcast is also called re-forecast, is the forecast of the past. E.g. you have a forecast from year 2000-2010, assuming now you are in 2005. So from 2000-2005, this period
is the hindcast period, and 2005-2010, this period is the forecast period.

Hindcast can be the same as forecast, i.e., you can use forecast itself as hindcast to train the bias correction.


\strong{How it works}

Forecast product has to be calibrated, usually the system is doing forecast in real time. So, e.g., if the
forecast starts from year 2000, assuming you are in year 2003, then you will have 3 years' hindcast
data (year 2000 - 2003), which can be used to calibrate. And your forecast period is (2003-2004)

E.g. you have observation from 2001 - 2002, this is your input obs. Then you can take the same
period (2001-2002) from the forecast, which is the hindcast period. For forecast, you can take any period.
The program will evaluate the obs and hindcast, to get the modification of the forecast, and then add the
modification to the forecast data.

\strong{method}

Different methods used in the bias correction.

\strong{delta}

This method consists on adding to the observations the mean change signal (delta method).
This method is applicable to any kind of variable but it is preferable to avoid it for bounded variables
 (e.g. precipitation, wind speed, etc.) because values out of the variable range could be obtained
 (e.g. negative wind speeds...)

 \strong{scaling}

This method consists on scaling the simulation  with the difference (additive) or quotient (multiplicative)
between the observed and simulated means in the train period. The \code{additive} or \code{multiplicative}
correction is defined by parameter \code{scaling.type} (default is \code{additive}).
The additive version is preferably applicable to unbounded variables (e.g. temperature)
and the multiplicative to variables with a lower bound (e.g. precipitation, because it also preserves the frequency).
}
\examples{
# Use testdl as an example, we take frc, hindcast and obs fro testdl.
data(testdl)

# common period has to be extracted in order to better train the forecast.

datalist <- extractPeriod(testdl, startDate = '1994-1-1', endDate = '1995-10-1')

frc <- datalist[[1]]
hindcast <- datalist[[2]]
obs <- datalist[[3]]

# default method is delta
frc_new <- biasCorrect(frc, hindcast, obs)

# If the variable is precipitation, it cannot be negative value, so use multi scale method
frc_new <- biasCorrect(frc, hindcast, obs, method = 'scaling', scaleType = 'multi')
}
\references{
Bias correction methods come from \code{biasCorrection} from \code{dowscaleR}

Santander Meteorology Group (2015). downscaleR: Climate data manipulation and statistical downscaling. R
package version 0.6-0. https://github.com/SantanderMetGroup/downscaleR/wiki
}

