% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{biasFactor-class}
\alias{biasFactor-class}
\title{An S4 class, representing the biasFactor of single time series biasCorrection.}
\description{
An S4 class, representing the biasFactor of single time series biasCorrection.
}
\section{Slots}{

\describe{
\item{\code{biasFactor}}{list of biasFactor, containing all the information for computing.}

\item{\code{method}}{the biascorrection method}

\item{\code{preci}}{if the data is precipitation}

\item{\code{scaleType}}{'Valid when 'scaling' method is selected, 'multi' or 'add'.}

\item{\code{extrapolate}}{Valid when 'eqm' method is selected, 'constant' or 'no'}

\item{\code{memberDim}}{members contained.}

\item{\code{prThreshold}}{precipitation threshold, under which the precipitation is considered as 0.}
}}

