% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{biasFactor.hyfo-class}
\alias{biasFactor.hyfo-class}
\title{An S4 class, representing the biasFactor of hyfo file.}
\description{
An S4 class, representing the biasFactor of hyfo file.
}
\section{Slots}{

\describe{
\item{\code{lonLatDim}}{lists of biasFactor}
}}

