% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/collectData_csv.R
\name{collectData_csv_anarbe}
\alias{collectData_csv_anarbe}
\title{Collect data from the gauging stations in spain, catchement Anarbe}
\usage{
collectData_csv_anarbe(folderName, output = T)
}
\arguments{
\item{folderName}{A string showing the path of the folder holding different csv files.}

\item{output}{A boolean showing whether the output is given, default is T.}
}
\value{
The collected data from different csv files.
}
\description{
Collect data from the gauging stations in spain, catchement Anarbe
}

