% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_anarbe.R
\name{collectData_excel_anarbe}
\alias{collectData_excel_anarbe}
\title{Collect data from different excel files}
\usage{
collectData_excel_anarbe(folderName, keyword = NULL, output = TRUE)
}
\arguments{
\item{folderName}{A string showing the folder path.}

\item{keyword}{A string showing the extracted column, e.g., waterLevel, waterBalance.}

\item{output}{A boolean showing whether the output is given.}
}
\value{
The collected data from different excel files.
}
\description{
Collect data from different excel files
}
\references{
\itemize{
\item R Core Team (2015). R: A language and environment for statistical computing. R Foundation for
Statistical Computing, Vienna, Austria. URL http://www.R-project.org/.
}
}
