% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord2cell}
\alias{coord2cell}
\title{Change lon lat coordinates to cell coordinates}
\usage{
coord2cell(coord, lon, lat)
}
\arguments{
\item{coord}{input lon lat coordinate}

\item{lon}{dataset lon array}

\item{lat}{dataset lat array}
}
\value{
A cell coordinate
}
\description{
Change lon lat coordinates to cell coordinates
}

