% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dataDocument.R
\docType{data}
\name{datalist}
\alias{datalist}
\title{datalist}
\format{A list consists of 3 different lists.
\describe{
  \item{AAA}{AAA, a dataframe containing a date column and a value column. }
  \item{BBB}{BBB, a dataframe containing a date column and a value column.}
  \item{CCC}{CCC, a dataframe containing a date column and a value column.}
  ...
}}
\usage{
datalist
}
\description{
A list containing different precipitation time series.
}
\keyword{datasets}

