% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/extractPeriod.R
\name{extractPeriod}
\alias{extractPeriod}
\title{Extract common period or certain period from a list of different dataframes of time series.}
\usage{
extractPeriod(datalist, startDate = NULL, endDate = NULL,
  commonPeriod = F)
}
\arguments{
\item{datalist}{A list of different dataframes of time series .}

\item{startDate}{A Date showing the start of the extract period, default as NULL.}

\item{endDate}{A Date showing the end of the extract period, default as NULL.}

\item{commonPeriod}{A boolean showing whether the common period is extracted. If chosen, startDate and endDate
should be NULL.}
}
\value{
A list with all the time series inside containing the same period.
}
\description{
Extract common period or certain period from a list of different dataframes of time series.
}
\references{
Data source:
http://meteo.navarra.es/estaciones/mapadeestaciones.cfm
http://www4.gipuzkoa.net/oohh/web/esp/02.asp
}

