% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/extractPeriod.R
\name{extractPeriod_dataframe}
\alias{extractPeriod_dataframe}
\title{Extract data from a dataframe with startDate and endDate}
\usage{
extractPeriod_dataframe(dataframe, startDate, endDate)
}
\arguments{
\item{dataframe}{A time series with first column being a series of date or time.}

\item{startDate}{A date representing the start date.}

\item{endDate}{A date representing the end date.}
}
\value{
The extracted dataframe between \code{startDate} and \code{endDate}.
}
\description{
Extract data from a dataframe with startDate and endDate
}

