% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fillGap.R
\name{fillGap}
\alias{fillGap}
\title{Fill gaps in the rainfall time series.}
\usage{
fillGap(dataset, corPeriod = "daily")
}
\arguments{
\item{dataset}{A dataframe with first column the time, the rest columns are rainfall data of different gauges}

\item{corPeriod}{A string showing the period used in the correlation computing, e.g. daily, monthly, annual.}
}
\value{
The filled dataframe
}
\description{
Fill gaps in the rainfall time series.
}
\examples{
b <- read.table(text = '        Date  AAA  BBB  CCC  DDD  EEE
49 1999-12-15 24.8 21.4 25.6 35.0 17.4
50 1999-12-16   NA  0.6  1.5  6.3  2.5
51 1999-12-17   NA 16.3 20.3  2.6 19.2
52 1999-12-18   NA  1.6 18.4  6.3  0.0
53 1999-12-19   NA 36.4 12.5 26.8 24.9
54 1999-12-20   NA  0.0  0.0  0.2  0.0
55 1999-12-21  0.2  0.0  0.0  0.0  0.0
56 1999-12-22  0.0  0.0  0.0  0.0  0.0')

b1 <- fillGap(b)
}

