% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getAnnual.R
\name{getAnnual_dataframe}
\alias{getAnnual_dataframe}
\title{Get annual rainfall of the input time series.}
\usage{
getAnnual_dataframe(dataset)
}
\arguments{
\item{dataset}{A dataframe containing one time series, e.g., rainfall from one gauging station.
the time should follow the format : "1990-1-1"}
}
\value{
The annual rainfall of each year of the input station.
}
\description{
Get annual rainfall of the input time series.
}

