% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getEnsemble.R
\name{getEnsem_comb}
\alias{getEnsem_comb}
\title{Combine ensembles together}
\usage{
getEnsem_comb(..., list = NULL, nrow = 1)
}
\arguments{
\item{...}{different ensembles generated by \code{getHisEnsem(, output = 'ggplot')}
or \code{getFrcEnsem(, output = 'ggplot')}, see details.}

\item{list}{If input is a list containing different ggplot data, use \code{list = inputlist}.}

\item{nrow}{A number showing the number of rows.}
}
\value{
A combined ensemble plot.
}
\description{
Combine ensembles together
}

