% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getEnsemble.R
\name{getFrcEnsem}
\alias{getFrcEnsem}
\title{getFrcEnsem extract different members' timeseries from forecasting data, if forecasting data has a member session.}
\usage{
getFrcEnsem(dataset, cell = "mean", plot = "norm", output = "data",
  name = NULL, mv = 0, ...)
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}, there should be a member part in the data part of the dataset.}

\item{cell}{A vector containing the locaton of the cell, e.g. c(2, 3), default is "mean", representing
the spatially averaged value. Check details for more information.}

\item{plot}{A string showing whether the plot will be shown, e.g., 'norm' means normal plot (without any process),
'cum' means cummulative plot, default is 'norm'. For other words there will be no plot.}

\item{output}{A string showing which type of output you want. Default is "data", if "ggplot", the
data that can be directly plotted by ggplot2 will be returned, which is easier for you to make series
plots afterwards. NOTE: If \code{output = 'ggplot'}, the missing value in the data will
be replaced by \code{mv}, if assigned, default mv is 0.}

\item{name}{A string showing the name of the ggplot output, in order to differentiate from other
ggplot output. Please see details. Only when \code{output = 'ggplot'}, name is valid, default is
the system time.}

\item{mv}{A number showing representing the missing value. When calculating the cumulative value,
missing value will be replaced by mv, default is 0.}

\item{...}{\code{title, x, y} showing the title and x and y axis of the plot. e.g. \code{title = 'aaa'}}
}
\value{
A ensemble time series extracted from forecating data.
}
\description{
getFrcEnsem extract different members' timeseries from forecasting data, if forecasting data has a member session.
}
\details{
\code{cell} representing the location of the cell, NOTE: this location means the index of the cell,
IT IS NOT THE LONGITUDE AND LATITUDE. e.g., \code{cell = c(2, 3)}, the program will take the 2nd longitude
and 3rd latitude, by the increasing order. Longitude comes first.

\code{name}
Assuming you have two ggplot outputs, you want to plot them together. In this situation, you
need a name column to differentiate one ggplot output from the other. You can assigne this name
by the argument directly, If name is not assigned and \code{output = 'ggplot'} is selected, then
the system time will be selected as name column.
}

