% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/analyzeTS.R
\name{getLMom}
\alias{getLMom}
\title{get L moment analysis of the input distribution}
\usage{
getLMom(dis)
}
\arguments{
\item{dis}{A distribution, for hydrology usually a time series with only data column without time.}
}
\value{
The mean, L-variation, L-skewness and L-kurtosis of the input distribution
}
\description{
get L moment analysis of the input distribution
}
\examples{
dis <- seq(1, 100)
getLMom(dis)
}
\references{
J. R. M. Hosking (2015). L-moments. R package, version 2.5. URL:
http://CRAN.R-project.org/package=lmom.
}

