% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getMeanPreci.R
\name{getMeanPreci}
\alias{getMeanPreci}
\title{Get mean rainfall data, e.g. mean annual rainfall, mean monthly rainfall and mean winter rainfall.}
\usage{
getMeanPreci(inputTS, method = NULL, yearIndex = NULL, monthIndex = NULL,
  fullResults = FALSE)
}
\arguments{
\item{inputTS}{A time series with only data.}

\item{method}{A string showing the method used to calculate mean value, e.g., "meanAnnualPreci".}

\item{yearIndex}{A list showing the year index of the time series.}

\item{monthIndex}{A list showing the month index of the time series.}

\item{fullResults}{A boolean showing whether the full results are shown.}
}
\value{
The mean value of the input time series or the full results before calculating mean.
}
\description{
Get mean rainfall data, e.g. mean annual rainfall, mean monthly rainfall and mean winter rainfall.
}

