% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getPreciBar.R
\name{getPreciBar}
\alias{getPreciBar}
\title{get mean rainfall bar plot of the input dataset}
\usage{
getPreciBar(dataset, method, outputData = T, plotRange = T)
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}.}

\item{method}{A string showing the calculating method of the input time series, e.g., "meanMonthly"}

\item{outputData}{A boolean showing whether the result will be returned, default is T.}

\item{plotRange}{A boolean showing whether the range will be plotted.}
}
\value{
The calculated mean value of the input time series and the plot of the result.
}
\description{
get mean rainfall bar plot of the input dataset
}

