% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getPreciBar.R
\name{getPreciBar_comb}
\alias{getPreciBar_comb}
\title{Combine bars together}
\usage{
getPreciBar_comb(..., list = NULL, nrow = 1)
}
\arguments{
\item{...}{different barplots generated by \code{getPreciBar(, output = 'ggplot')}}

\item{list}{If input is a list containing different ggplot data, use l\code{list = inputlist}.}

\item{nrow}{A number showing the number of rows.}
}
\value{
A combined barplot.
}
\description{
Combine bars together
}

