% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSpatialMap.R
\name{getSpatialMap}
\alias{getSpatialMap}
\title{Get spatial map of the input dataset.}
\usage{
getSpatialMap(dataset, method = NULL, ...)
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}.}

\item{method}{A string showing different calculating method for the map.}

\item{...}{\code{title, y} showing the title and x and y axis of the plot, default is about precipitation.}
}
\value{
A matrix representing the raster map is returned, and the map is plotted.
}
\description{
Get spatial map of the input dataset.
}

