% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSpatialMap.R
\name{getSpatialMap_mat}
\alias{getSpatialMap_mat}
\title{Get spatial map of the input dataset, and a matrix representing the raster map will be returned.}
\usage{
getSpatialMap_mat(matrix, title = NULL, catchment = NULL, points = NULL,
  output = "data", info = TRUE, scale = "identity", ...)
}
\arguments{
\item{matrix}{A matrix raster, should be the result of \code{getSpatialMap()}, output should be default
or 'data'}

\item{title}{A string showing the title of the plot, defaut is NULL.}

\item{catchment}{A catchment file geting from \code{shp2cat()} in the package, if a catchment is available for background.}

\item{points}{A dataframe, showing other information, e.g., location of the gauging stations. The
the data.frame should be with columes "name, lon, lat, z, value".}

\item{output}{A string showing the type of the output, if \code{output = 'ggplot'}, the returned
data can be used in ggplot and \code{getSpatialMap_comb()}; if \code{output = 'plot'}, the returned data is the plot containing all
layers' information, and can be plot directly or used in grid.arrange; if not set, the raster matrix data
will be returned.}

\item{info}{A boolean showing whether the information of the map, e.g., max, mean ..., default is T.}

\item{scale}{A string showing the plot scale, 'identity' or 'sqrt'.}

\item{...}{\code{title, y} showing the title and x and y axis of the plot, default is about precipitation.}
}
\value{
A matrix representing the raster map is returned, and the map is plotted.
}
\description{
Get spatial map of the input dataset, and a matrix representing the raster map will be returned.
}
\examples{
data(tgridData)# the result of \\code{loadGridData{ecomsUDG.Raccess}}
#the output type of has to be default or 'data'.
a1 <- getSpatialMap(tgridData, method = 'mean')
a2 <- getSpatialMap(tgridData, method = 'max')
a3 <- getSpatialMap(tgridData, method = 'winter')
a4 <- getSpatialMap(tgridData, method = 'summer')
#For example, if we want to investigate the difference between mean value and max.

a5 <- a2 - a1
getSpatialMap_mat(a4)

#Or to investigate the difference between winter value and summer value.
a6 <- a3 - a4
getSpatialMap_mat(a6)
}

