% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getDataRange.R
\name{getSpatialRange}
\alias{getSpatialRange}
\title{Get spatial longitude and latitude range of the input netcdf file}
\usage{
getSpatialRange(filePath)
}
\arguments{
\item{filePath}{A string showing the path of the netcdf file.}
}
\value{
A vector of the range of the longitude and latitude.
}
\description{
Get spatial longitude and latitude range of the input netcdf file
}

