% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/list2dataframe.R
\name{list2Dataframe}
\alias{list2Dataframe}
\title{Convert a list of different time series to a dataframe}
\usage{
list2Dataframe(datalist)
}
\arguments{
\item{datalist}{A list containing different time series, each sub list has to have the same length.}
}
\value{
The converted dataframe
}
\description{
Convert a list of different time series to a dataframe
}

