% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/monDay.R
\name{monDay}
\alias{monDay}
\title{Monthly data to daily data and the reverse conversion.}
\usage{
monDay(TS, method)
}
\arguments{
\item{TS}{A time series, with first column date, and second column value. The date column should
follow the format in \code{as.Date}, i.e. seperate with "-" or "/". Check details for more information.}

\item{method}{A string showing whether you want to change a daily data to monthly data or monthly
data to daily data.e.g. "mon2day" and "day2mon".}
}
\value{
converted time series.
}
\description{
Monthly data to daily data and the reverse conversion.
}
\details{
Note, when you want to change daily data to monthly data, a new date column will be generated,
usually the date column will be the middle date of each month, 15th, or 16th. However, if your
time series doesn't start from the beginning of a month or ends to the end of a month, e.g.
from 1999-3-14 to 2008-2-2, the first and last generated date could be wrong. Not only the date, but also the data, because you are
not calculating based on a intact month.
}
\examples{
# Daily to monthly
data(testdl)
TS <- testdl[[2]] # Get daily data
str(TS)
TS_new <- monDay(TS, method = 'day2mon')

# Monthly to daily
TS <- data.frame(Date = seq(as.Date('1999-9-15'), length = 30, by = '1 month'),
runif(30, 3, 10))
TS_new <- monDay(TS, method = 'mon2day')

# More examples can be found in the user manual on http://yuanchao-xu.github.io/hyfo/
}
\references{
R Core Team (2015). R: A language and environment for statistical computing. R Foundation for
Statistical Computing, Vienna, Austria. URL http://www.R-project.org/.
}

