% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGap.R
\name{monthlyPreci}
\alias{monthlyPreci}
\title{Get monthly rainfall}
\usage{
monthlyPreci(TS, year, mon)
}
\arguments{
\item{TS}{A rainfall time series.}

\item{year}{A list showing the year index of the time series.}

\item{mon}{A list showing the mon index of the time series.}
}
\value{
the monthly rainfall matrix of the rainfall time series.
}
\description{
Get monthly rainfall
}
