% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/analyzeTS.R
\name{plotTS_comb}
\alias{plotTS_comb}
\title{Combine time seires plot together}
\usage{
plotTS_comb(..., nrow = 1, type = "line", list = NULL, x = "Date",
  y = "", title = "", output = FALSE)
}
\arguments{
\item{...}{different time series plots generated by \code{plotTS(, output = 'ggplot')}, refer to details.}

\item{nrow}{A number showing the number of rows.}

\item{type}{A string showing 'line' or 'bar'.}

\item{list}{If input is a list containing different ggplot data, use l\code{list = inputlist}.}

\item{x}{A string of x axis name.}

\item{y}{A string of y axis name.}

\item{title}{A string of the title.}

\item{output}{A boolean, if chosen TRUE, the output will be given.
NOTE: yOU HAVE TO PUT A \code{list = }, before your list.}
}
\value{
A combined time series plot.
}
\description{
Combine time seires plot together
}
\details{
..., representing different ouput generated by \code{plotTS(, output = 'ggplot')}, different
names must be assigned when generating different output.
}
\examples{
# comming soon.
}
\references{
H. Wickham. ggplot2: elegant graphics for data analysis. Springer New York, 2009.
}

