% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/shp2cat.R
\name{shp2cat}
\alias{shp2cat}
\title{Get a catchment object from selected shape file.
Run the function, and a window will open up to choose shape file.}
\usage{
shp2cat()
}
\value{
A catchment object can be used in \code{getSpatialMap()}.
}
\description{
Get a catchment object from selected shape file.
Run the function, and a window will open up to choose shape file.
}

