% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocument.R
\docType{data}
\name{testCat}
\alias{testCat}
\title{testCat}
\format{A catchment file generated by library rgdal.
\describe{
  \item{class}{Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots}
  ...
}}
\usage{
testCat
}
\description{
testCat
}
\keyword{datasets}

