% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocument.R
\docType{data}
\name{testdl}
\alias{testdl}
\title{testdl}
\format{A list consists of 3 different lists.
\describe{
  \item{AAA}{AAA, a dataframe containing a date column and a value column. }
  \item{BBB}{BBB, a dataframe containing a date column and a value column.}
  \item{CCC}{CCC, a dataframe containing a date column and a value column.}
  ...
}}
\source{
http://meteo.navarra.es/estaciones/mapadeestaciones.cfm
http://www4.gipuzkoa.net/oohh/web/esp/02.asp
}
\usage{
testdl
}
\description{
A list containing different precipitation time series.
}
\references{
\itemize{
\item http://meteo.navarra.es/estaciones/mapadeestaciones.cfm
\item #' http://www4.gipuzkoa.net/oohh/web/esp/02.asp
}
}
\keyword{datasets}

