% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dataDocument.R
\docType{data}
\name{tgridData}
\alias{tgridData}
\title{tgridData}
\format{A list containing different information.
\describe{
  \item{Variables}{variable information. }
  \item{Data}{Data.}
  \item{xyCoords}{longitude and latitude of the data.}
  \item{Dates}{Date information.}
  ...
}}
\source{
http://www.meteo.unican.es/datasets/spain02
}
\usage{
tgridData
}
\description{
A list containing different information getting from grid data file, e.g., netcdf file.
}
\references{
Herrera, S., Ancell, R., Gutierrez, J. M., Pons, M. R., Frias, M. D., & Fernandez, J.
(2012). Development and analysis of a 50-year high-resolution daily gridded precipitation dataset
over Spain (Spain02). International Journal of Climatology
(http://www.meteo.unican.es/datasets/spain02), 10.1002/joc.2256.
}
\keyword{datasets}

