% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/collectData.R
\name{collectData}
\alias{collectData}
\title{Collect data from different csv files.}
\usage{
collectData(folderName, fileType = NULL, range = NULL, sheetIndex = 1)
}
\arguments{
\item{folderName}{A string showing the path of the folder holding different csv files.}

\item{fileType}{A string showing the file type, e.g. "txt", "csv", "excel".}

\item{range}{A vector containing startRow, endRow, startColumn, endColumn, e.g.,
c(2,15,2,3)}

\item{sheetIndex}{A number showing the sheetIndex in the excel file, if fileType is excel,
sheetIndex has to be provided, default is 1.}
}
\value{
The collected data from different files in the folder.
}
\description{
Collect data from different csv files.
}
\examples{
#use internal data as an example.
file <- system.file("extdata", "1999.csv", package = "hyfo")
folder <- strsplit(file, '1999')[[1]][1]
a <- collectData(folder, fileType = 'csv', range = c(10, 20, 1,2))
}

