% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/collectData_txt.R
\name{collectData_txt_anarbe}
\alias{collectData_txt_anarbe}
\title{collect data from different txt.}
\usage{
collectData_txt_anarbe(folderName, output = T, rangeWord = c("Ene       ",
  -1, "Total     ", -6))
}
\arguments{
\item{folderName}{A string showing the folder path.}

\item{output}{A boolean showing whether the result is given.}

\item{rangeWord}{A list containing the keyword and the shift.
defaut is set to be used in spain gauging station.}
}
\value{
The collected data from different txt files.
}
\description{
collect data from different txt.
}

