% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/extractPeriod.R
\name{extractPeriod}
\alias{extractPeriod}
\title{Extract period from list or dataframe.}
\usage{
extractPeriod(datalist, startDate = NULL, endDate = NULL,
  commonPeriod = FALSE, dataframe = NULL)
}
\arguments{
\item{datalist}{A list of different dataframes of time series .}

\item{startDate}{A Date showing the start of the extract period, default as NULL.}

\item{endDate}{A Date showing the end of the extract period, default as NULL.}

\item{commonPeriod}{A boolean showing whether the common period is extracted. If chosen, startDate and endDate
should be NULL.}

\item{dataframe}{A dataframe with first column Date, the rest columns value. If your input is a
dataframe, not time series list, you can put \code{dataframe = yourdataframe}. And certain period will be
extracted. Note: if your input is a time series, that means all the columns share the same period of date.}
}
\value{
A list or a dataframe with all the time series inside containing the same period.
}
\description{
Extract common period or certain period from a list of different dataframes of time series, or from a
dataframe.
NOTE: all the dates in the datalist should follow the format in ?as.Date{base}.
}
\examples{
# Generate timeseries datalist. Each data frame consists of a Date and a value.

AAA <- data.frame(
# date column
Date = seq(as.Date('1990-10-28'),as.Date('1997-4-1'),1),
 # value column
AAA = sample(1:100,length(seq(as.Date('1990-10-28'),as.Date('1997-4-1'),1)), repl = TRUE))

BBB <- data.frame(
Date = seq(as.Date('1993-3-28'),as.Date('1999-1-1'),1),
BBB = sample(1:100,length(seq(as.Date('1993-3-28'),as.Date('1999-1-1'),1)), repl = TRUE))

CCC <- data.frame(
Date = seq(as.Date('1988-2-2'),as.Date('1996-1-1'),1),
CCC = sample(1:100,length(seq(as.Date('1988-2-2'),as.Date('1996-1-1'),1)), repl = TRUE))

list <- list(AAA, BBB, CCC)# dput() and dget() can be used to save and load list file.

list_com <- extractPeriod(list, commonPeriod = TRUE)

# list_com is the extracted datalist.
str(list_com)

# If startDate and endDate is provided, the record between them will be extracted.
# make sure startDate is later than any startDate in each dataframe and endDate is
# earlier than any endDate in each dataframe.

data(testdl)
datalist_com1 <- extractPeriod(testdl, startDate = '1994-1-1', endDate = '1995-10-1')
}
\references{
Achim Zeileis and Gabor Grothendieck (2005). zoo: S3 Infrastructure for Regular and Irregular Time
Series. Journal of Statistical Software, 14(6), 1-27. URL http://www.jstatsoft.org/v14/i06/
}

