% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getAnnual.R
\name{getAnnual}
\alias{getAnnual}
\title{Get annual rainfall of different rainfall time series}
\usage{
getAnnual(datalist)
}
\arguments{
\item{datalist}{A list containing different time series of different rainfall gauges}
}
\value{
The annual rainfall and the number of missing data of each year and each rainfall gauge
}
\description{
Get annual rainfall of different rainfall time series
}
\references{
Data source:
http://meteo.navarra.es/estaciones/mapadeestaciones.cfm
http://www4.gipuzkoa.net/oohh/web/esp/02.asp
}

