% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getAnnual.R
\name{getAnnual}
\alias{getAnnual}
\title{Get annual rainfall of different rainfall time series}
\usage{
getAnnual(datalist, output = "series", minRecords = 355, ...)
}
\arguments{
\item{datalist}{A list containing different time series of different rainfall gauges.}

\item{output}{A string showing the output output.}

\item{minRecords}{A number showing the minimum accept record number, e.g. for a normal
year(365 days), if \code{minRecords = 360}, it means if a year has less than 360 records
of a year, it will be ignored in the mean annual value calculation. Only valid
when \code{output = "mean"}, default is 355.}

\item{...}{\code{title, x, y} showing the title and x and y axis of the plot.}
}
\value{
The annual rainfall and the number of missing data of each year and each rainfall gauge, which
will also be plotted. If output "mean" is seleted, the mean annual rainfall will be returned.
}
\description{
Get annual rainfall of different rainfall time series
}
\examples{
#datalist is provided by the package as a test.
data(testdl)
a <- getAnnual(testdl)
#set minRecords to control the calculation of annual rainfall.
b <- getAnnual(testdl, output = 'mean', minRecords = 350)
c <- getAnnual(testdl, output = 'mean', minRecords = 365)
}

