% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getEnsemble.R
\name{getEnsem_comb}
\alias{getEnsem_comb}
\title{Combine ensembles together}
\usage{
getEnsem_comb(..., list = NULL, nrow = 1, legend = TRUE, x = "",
  y = "", title = "", output = FALSE)
}
\arguments{
\item{...}{different ensembles generated by \code{getHisEnsem(, output = 'ggplot')}
or \code{getFrcEnsem(, output = 'ggplot')}, see details.}

\item{list}{If input is a list containing different ggplot data, use \code{list = inputlist}.}

\item{nrow}{A number showing the number of rows.}

\item{legend}{A boolean representing whether you want the legend. Sometimes when you combine
plots, there will be a lot of legends, if you don't like it, you can turn it off by setting
\code{legend = FALSE}, default is TRUE.}

\item{x}{A string of x axis name.}

\item{y}{A string of y axis name.}

\item{title}{A string of the title.}

\item{output}{A boolean, if chosen TRUE, the output will be given.}
}
\value{
A combined ensemble plot.
}
\description{
Combine ensembles together
}
\references{
H. Wickham. ggplot2: elegant graphics for data analysis. Springer New York, 2009.
}

