% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getEnsemble.R
\name{getFrcEnsem}
\alias{getFrcEnsem}
\title{getFrcEnsem extract different members' timeseries from forecasting data, if forecasting data has a member session.}
\usage{
getFrcEnsem(dataset, cell = "mean", plot = TRUE)
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}, there should be a member part in the data part of the dataset.}

\item{cell}{A vector containing the locaton of the cell, e.g. c(2, 3), default is "mean", representing
the spatially averaged value. Check details for more information.}

\item{plot}{A boolean showing whether the plot will be shown, default is TRUE.}
}
\value{
A ensemble time series extracted from forecating data.
}
\description{
getFrcEnsem extract different members' timeseries from forecasting data, if forecasting data has a member session.
}
\details{
\code{cell} representing the location of the cell, NOTE: this location means the index of the cell,
IT IS NOT THE LONGITUDE AND LATITUDE. e.g., \code{cell = c(2, 3)}, the program will take the 2nd longitude
and 3rd latitude, by the increasing order. Longitude comes first.
}

