% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSpatialMap.R
\name{getSpatialMap}
\alias{getSpatialMap}
\title{Get spatial map of the input dataset, and a matrix representing the raster map will be returned.}
\usage{
getSpatialMap(dataset, method = NULL, ...)
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}.}

\item{method}{A string showing different calculating method for the map.}

\item{...}{\code{title, y} showing the title and x and y axis of the plot, default is about precipitation.
\code{limits}, \code{breaks}, see \code{scale_fill_gradientn()} for more details.}
}
\value{
A matrix representing the raster map is returned, and the map is plotted.
}
\description{
Get spatial map of the input dataset, and a matrix representing the raster map will be returned.
}

