% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSpatialMap.R
\name{getSpatialMap}
\alias{getSpatialMap}
\title{Get spatial map of the input dataset.}
\usage{
getSpatialMap(dataset, method = NULL, ...)
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}.}

\item{method}{A string showing different calculating method for the map.}

\item{...}{Check \code{?getSpatialMap_mat} for details, e.g., x, y, title, catchment,
points, output,}
}
\value{
A matrix representing the raster map is returned, and the map is plotted.
}
\description{
Get spatial map of the input dataset.
}
\examples{
data(gridData) # the result of \\code {loadGridData{ecomsUDG.Raccess}}
getSpatialMap(gridData, method = 'meanAnnual')
getSpatialMap(gridData, method = 'winter')

data(testCat)
getSpatialMap(gridData, method = 'winter', catchment = testCat)

file <- system.file("extdata", "points.txt", package = "hyfo")
points <- read.table(file, header = TRUE, sep = ',' )
getSpatialMap(gridData, method = 'winter', catchment = testCat, points = points)
}

