% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSpatialMap.R
\name{getSpatialMap}
\alias{getSpatialMap}
\title{Get spatial map of the input dataset, and a matrix representing the raster map will be returned.}
\usage{
getSpatialMap(dataset, catchment = NULL, points = NULL, method = NULL,
  outputData = TRUE, plotScale = "identity")
}
\arguments{
\item{dataset}{A list containing different information, should be the result of reading netcdf file using
\code{library(ecomsUDG.Raccess)}.}

\item{catchment}{A catchment file geting from \code{shp2cat()} in the package, if a catchment is available for background.}

\item{points}{A shape file showing other information, e.g., location of the gauging stations.}

\item{method}{A string showing different calculating method for the map.}

\item{outputData}{A boolean showing whether the raster matrix will be returned, default is T.}

\item{plotScale}{A string showing the plot scale, e.g., "identity" or "sqrt"}
}
\value{
A matrix representing the raster map is returned, and the map is plotted.
}
\description{
Get spatial map of the input dataset, and a matrix representing the raster map will be returned.
}

