% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSpatialMap.R
\name{getSpatialMap_comb}
\alias{getSpatialMap_comb}
\title{Combine maps together}
\usage{
getSpatialMap_comb(..., list = NULL, nrow = 1)
}
\arguments{
\item{...}{different maps generated by \code{getSpatialMap(, output = 'ggplot')}}

\item{list}{If input is a list containing different ggplot data, use l\code{list = inputlist}.}

\item{nrow}{A number showing the number of rows.}
}
\value{
A combined map.
}
\description{
Combine maps together
}
\examples{
data(tgridData)# the result of \\code{loadGridData{ecomsUDG.Raccess}}
#The output should be 'ggplot'
a1 <- getSpatialMap(tgridData, method = 'summer', output = 'ggplot')
a2 <- getSpatialMap(tgridData, method = 'winter', output = 'ggplot')
a3 <- getSpatialMap(tgridData, method = 'mean', output = 'ggplot')
a4 <- getSpatialMap(tgridData, method = 'max', output = 'ggplot')
getSpatialMap_comb(a1, a2, a3, a4)
getSpatialMap_comb(a1, a2, a3, a4, nrow = 2)
}

