% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Ncdf_related.R
\name{loadNcdf}
\alias{loadNcdf}
\title{Load NetCDF file}
\usage{
loadNcdf(filePath, varname, tz = "GMT")
}
\arguments{
\item{filePath}{A path pointing to the NetCDF file, version3.}

\item{varname}{A character representing the variable name, you can use \code{getNcdfVar} to
get the basic information about the variables and select the target.}

\item{tz}{A string representing the time zone, default is GMT, if you know what time zone is
you can assign it in the argument. If \code{tz = ''}, current time zone will be taken.}
}
\value{
A list object from \code{hyfo} containing the information to be used in the analysis,
or biascorrection.
}
\description{
Load NetCDF file
}

