% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/analyzeTS.R
\name{plotTS}
\alias{plotTS}
\title{plot time series, with marks on missing value.}
\usage{
plotTS(TS, type = "line", output = "data", name = NULL, ...)
}
\arguments{
\item{TS}{A time series in form of a dataframe.}

\item{type}{A string representing the type of the time series, e.g. 'line' or 'bar'.}

\item{output}{A string showing which type of output you want. Default is "data", if "ggplot", the
data that can be directly plotted by ggplot2 will be returned, which is easier for you to make series
plots afterwards.}

\item{name}{If \code{output = 'ggplot'}, name has to be assigned to your output, in order to differentiate
different outputs in the later multiplot using \code{plotTS_comb}.}

\item{...}{\code{title, x, y} showing the title and x and y axis of the plot. e.g. \code{title = 'aaa'},
\code{x = 'aaa'}, \code{y = 'aaa'}}
}
\value{
A plot of the input time series.
}
\description{
plot time series, with marks on missing value.
}
\examples{
plotTS(testdl[[1]])
plotTS(testdl[[1]], x = 'xxx', y = 'yyy', title = 'aaa')


# More examples can be found in the user manual on http://yuanchao-xu.github.io/hyfo/
}
\references{
H. Wickham. ggplot2: elegant graphics for data analysis. Springer New York, 2009.
}

