% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/shp2cat.R
\name{shp2cat}
\alias{shp2cat}
\title{Get a catchment object from selected shape file.}
\usage{
shp2cat(filePath)
}
\arguments{
\item{filePath}{A string representing the path of the shape file.}
}
\value{
A catchment object can be used in \code{getSpatialMap()}.
}
\description{
Get a catchment object from selected shape file.
}
\details{
This function is based on the package \code{rgdal} and \code{sp}, and the output comes from the package
\code{sp}
}
\examples{
#open internal file
file <- system.file("extdata", "testCat.shp", package = "hyfo")
catchment <- shp2cat(file)
}
\references{
rgdal package and the
}

